Move ConfigForm & theme to new separate files; Create SmallIconNavButton and SmallIconButton to avoid code duplication
